requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter
alembic==1.13.0

# Security
Werkzeug==3.0.1
python-dotenv==1.0.0
bcrypt==4.1.2

# AI Services
google-generativeai==0.3.2
elevenlabs==2.16.0
manim==0.18.1
requests==2.32.3
python-docx==1.1.2
pypdf==4.0.1
python-pptx==0.6.23

# Utilities
redis==5.0.1  # For rate limiting in production
gunicorn==21.2.0  # Production WSGI server
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
coverage==7.3.2

# Code quality
flake8==6.1.0
black==23.12.1

# Development
watchdog==3.0.0
